chore: moved cypress into test folder
