--- cypress/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"noEmit": true,
-		"types": ["node", "cypress", "cypress-real-events", "@testing-library/cypress"]
-	},
-	"include": ["../cypress.d.ts", "**/*.js", "**/*.jsx", "**/*.tsx", "**/*.ts"]
-}
