+++ test/cypress/component-testing/components/keyboard-only/index.spec.tsx
