+++ test/cypress/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"sourceMap": false,
+		"noEmit": true,
+		"types": ["node", "cypress", "cypress-real-events", "@testing-library/cypress"]
+	},
+	"include": ["../cypress.d.ts", "**/*.js", "**/*.jsx", "**/*.tsx", "**/*.ts"]
+}
